Move `conda/shell` tests into new `tests/test_shell.py` (#14141)
* Move shell tests into separate test file

* Move --dev comment to the more relevant location

* Reorder logging/imports

* Remove unused PYTHONIOENCODING

* Mark entire test_shell.py as integration

* Remove test_unicode (for Python 2)

* Replace capture context with capsys

* Flatten string formatting

* Move stacking test to test_shell

* Fix test_stacking

* Drop --dev from test_activate.py

* Replace open with Path

* Replace tempdir with tmp_path

* Remove dals

* Replace mkdir_p & touch with Path
